omg hw6_ex2 works!!!
The issue that has been plaguing me for the last hour and a half today
was that when I was rolling the bits to the left, I had the number in
ebx instead of bx, so when I thought I should be rotating a 1 off the
end, it was actually rotating a 0 because there was the 16 0s in the
front that I wasn't expecting. Such a stupid mistake.